requirements.txt
# fabric
fabric3

dropbox
PyYAML
tqdm
